pyproject.toml
[tool.poetry]
name = "llm-study"
version = "0.1.0"
description = ""
authors = ["ding <email81227@gmail.com>"]
readme = "README.md"
packages = [{include = "llm_study"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
urllib3 = "1.26.15"
langchain = "0.1.8"
chainlit = "^1.0.200"
openai = "1.12"
langfuse = "^2.13.4"
pydantic-settings = "^2.2.0"
plotly = "^5.19.0"
langchain-experimental = "^0.0.52"
langchain-openai = "^0.0.8"
pandas = "^2.2.1"
pdfplumber = "0.10.2"
transformers = "^4.38.2"
sentence-transformers = "^2.5.1"
yfinance = "^0.2.37"
finlab = "^1.0.9"
faiss-cpu = "^1.8.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
